Fix impossible reference arguments
The way threaded_image_transform is implemented, there is no way it can
pass an argument to the transform function by reference in any
meaningful way. Fortunately, since the arguments are vil_image_view's,
it also doesn't need to; a copy of the view will still have a reference
to the underlying storage. Accordingly, remove impossible reference
qualifications from vil_image_view parameters meant to convey output
buffers.

This appears as a build error with newer versions of Boost, which are
stricter about this sort of nonsense.
(See also https://stackoverflow.com/questions/45549724.)